requirements.txt
fastmcp
faiss-cpu
sentence-transformers
numpy
beautifulsoup4
requests
python-dotenv
httpx
openai
